move rest to http plugin
