+++ jrcc-access-spring-boot-autoconfigure/src/main/java/ca/gov/bc/open/jrccaccess/autoconfigure/plugins/http/AutoConfiguration.java
@@ -1,4 +1,4 @@
-package ca.gov.bc.open.jrccaccess.autoconfigure;
+package ca.gov.bc.open.jrccaccess.autoconfigure.plugins.http;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.ComponentScan;
@@ -10,7 +10,7 @@ import org.springframework.context.annotation.Configuration;
  * @since 0.2.0
  */
 @Configuration
-@ComponentScan("ca.gov.bc.open.jrccaccess.autoconfigure.rest")
-public class RestAutoConfiguration {
+@ComponentScan
+public class AutoConfiguration {
 
 }
