+++ jrcc-access-spring-boot-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -2,4 +2,4 @@ org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 ca.gov.bc.open.jrccaccess.autoconfigure.AccessAutoConfiguration,\
 ca.gov.bc.open.jrccaccess.autoconfigure.plugins.rabbitmq.AutoConfiguration,\
 ca.gov.bc.open.jrccaccess.autoconfigure.plugins.console.AutoConfiguration,\
-ca.gov.bc.open.jrccaccess.autoconfigure.RestAutoConfiguration
+ca.gov.bc.open.jrccaccess.autoconfigure.plugins.http.AutoConfiguration
